Level: Moved LightPopulationTask out of generator namespace
